\name{semiaftosmac4-package}
\alias{semiaftosmac4-package}
\alias{semiaftosmac4}
\docType{package}
\title{
\packageTitle{semiaftosmac4}
}
\description{
\packageDescription{semiaftosmac4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{semiaftosmac4}
\packageIndices{semiaftosmac4}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{semiaftosmac4}

Maintainer: \packageMaintainer{semiaftosmac4}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
